+++ yggdrasil_decision_forests/port/python/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## HEAD
+
+-   Added model self evaluation.
+
 ## 0.0.8 - 2023-12-19
 
 ### Features